+++ doc/changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+# 1.2.16 - 2017-09-04
+### Added
+* underline and justify in tynemce
+* post type added to main query for custom taxonomy archive page
+
 # 1.2.15 - 2017-08-31
 ### Fixed
 * wp_seo taxonomy bug